gemini/requirements.txt
# 推荐的 Python 版本: 3.11 或 3.12
google-generativeai
pandas
tqdm
scikit-learn
seaborn
matplotlib
errant
spacy

# (仍然需要) 单独下载 spaCy 模型
# python -m spacy download en_core_web_sm